+++ NQueue/Internal/Model/WorkItemInfo.cs
@@ -1,6 +1,6 @@
 ﻿namespace NQueue.Internal.Model
 {
 
-    internal record WorkItemInfo(int WorkItemId, string Url, string? Internal);
+    internal record WorkItemInfo(long WorkItemId, string Url, string? Internal);
 
 }
